fix: move meal plan data fetching to CalendarPage
